+++ test/bract/cli/inducer_test.clj
@@ -11,7 +11,7 @@
   (:require
     [clojure.test :refer :all]
     [bract.core.keydef :as core-kdef]
-    [bract.cli.config  :as clim-config]
+    [bract.cli.keydef  :as clim-kdef]
     [bract.cli.inducer :as clim-inducer])
   (:import
     [java.io ByteArrayInputStream StringReader]))
@@ -21,44 +21,44 @@
   (is (thrown? IllegalArgumentException
         (clim-inducer/parse-args {})) "missing CLI args")
   (is (reduced?
-        (clim-inducer/parse-args {(key clim-config/ctx-config-required?) true
+        (clim-inducer/parse-args {(key clim-kdef/ctx-config-required?) true
                                   (key core-kdef/ctx-cli-args) ["-c" "hey"]})) "config file required but not passed")
   (let [context {(key core-kdef/ctx-cli-args) ["-f" "foo.edn"]}]
     (is (= (assoc context
-             (key clim-config/ctx-command) "run"
-             (key clim-config/ctx-cmd-args) []
+             (key clim-kdef/ctx-command) "run"
+             (key clim-kdef/ctx-cmd-args) []
              (key core-kdef/ctx-verbose?) false
              (key core-kdef/ctx-config-files) ["foo.edn"])
           (clim-inducer/parse-args context)) "specified config file"))
   (let [context {(key core-kdef/ctx-cli-args) ["-f" "foo.edn,bar.edn"]}]
     (is (= (assoc context
-             (key clim-config/ctx-command) "run"
-             (key clim-config/ctx-cmd-args) []
+             (key clim-kdef/ctx-command) "run"
+             (key clim-kdef/ctx-cmd-args) []
              (key core-kdef/ctx-verbose?) false
              (key core-kdef/ctx-config-files) ["foo.edn" "bar.edn"])
           (clim-inducer/parse-args context)) "specified multiple config files"))
   (let [context {(key core-kdef/ctx-cli-args) ["-c" "hey"]}]
     (is (= (assoc context
-             (key clim-config/ctx-command) "hey"
-             (key clim-config/ctx-cmd-args) []
+             (key clim-kdef/ctx-command) "hey"
+             (key clim-kdef/ctx-cmd-args) []
              (key core-kdef/ctx-verbose?) false)
           (clim-inducer/parse-args context)) "specified command"))
   (let [context {(key core-kdef/ctx-cli-args) ["-v"]}]
     (is (= (assoc context
              (key core-kdef/ctx-verbose?) true
-             (key clim-config/ctx-command) "run"
-             (key clim-config/ctx-cmd-args) [])
+             (key clim-kdef/ctx-command) "run"
+             (key clim-kdef/ctx-cmd-args) [])
           (clim-inducer/parse-args context)) "specified verbosity flag"))
   (let [context {(key core-kdef/ctx-cli-args) ["-vc" "hey"]}]
     (is (= (assoc context
-             (key clim-config/ctx-command) "hey"
-             (key clim-config/ctx-cmd-args) []
+             (key clim-kdef/ctx-command) "hey"
+             (key clim-kdef/ctx-cmd-args) []
              (key core-kdef/ctx-verbose?) true)
           (clim-inducer/parse-args context)) "specified command and verbose"))
   (let [context {(key core-kdef/ctx-cli-args) ["-vc" "hey" "foo" "bar"]}]
     (is (= (assoc context
-             (key clim-config/ctx-command) "hey"
-             (key clim-config/ctx-cmd-args) ["foo" "bar"]
+             (key clim-kdef/ctx-command) "hey"
+             (key clim-kdef/ctx-cmd-args) ["foo" "bar"]
              (key core-kdef/ctx-verbose?) true)
           (clim-inducer/parse-args context)) "specified verbose, command and arguments")))
 
@@ -67,20 +67,20 @@
   (is (thrown? IllegalArgumentException
         (clim-inducer/execute-command {})) "missing command")
   (is (thrown? IllegalArgumentException
-        (clim-inducer/execute-command {(key clim-config/ctx-command) "run"})) "missing command args")
+        (clim-inducer/execute-command {(key clim-kdef/ctx-command) "run"})) "missing command args")
   (is (not (reduced?
-             (clim-inducer/execute-command {(key clim-config/ctx-command) "run"
-                                            (key clim-config/ctx-cmd-args) []}))) "the `run` command")
+             (clim-inducer/execute-command {(key clim-kdef/ctx-command) "run"
+                                            (key clim-kdef/ctx-cmd-args) []}))) "the `run` command")
   (is (not (reduced?
-             (clim-inducer/execute-command {(key clim-config/ctx-command) "dryrun"
-                                            (key clim-config/ctx-cmd-args) []}))) "the `dryrun` command")
+             (clim-inducer/execute-command {(key clim-kdef/ctx-command) "dryrun"
+                                            (key clim-kdef/ctx-cmd-args) []}))) "the `dryrun` command")
   (is (reduced?
         (clim-inducer/execute-command {(key core-kdef/ctx-config) {"foo" "bar"}
-                                       (key clim-config/ctx-command) "config"
-                                       (key clim-config/ctx-cmd-args) []})) "the `config` command")
+                                       (key clim-kdef/ctx-command) "config"
+                                       (key clim-kdef/ctx-cmd-args) []})) "the `config` command")
   (comment  ; this particular test/assertion messes with test reports
     (is (reduced?
           (binding [*in* (StringReader. "")]
             (clim-inducer/execute-command {(key core-kdef/ctx-config) {"foo" "bar"}
-                                           (key clim-config/ctx-command) "repl"
-                                           (key clim-config/ctx-cmd-args) []}))) "the `repl` command")))
+                                           (key clim-kdef/ctx-command) "repl"
+                                           (key clim-kdef/ctx-cmd-args) []}))) "the `repl` command")))